feat: expose shared selectors, mutators, and root state
