+++ src/features/adventure/ui/zoneUI.js
@@ -1,4 +1,4 @@
-import { S } from '../../../game/state.js';
+import { S } from '../../../shared/state.js';
 import { ZONES } from '../data/zones.js';
 
 export function updateZoneButtons(selectZone) {
