+++ src/features/inventory/ui/CharacterPanel.js
@@ -1,4 +1,4 @@
-import { S, save } from '../../../game/state.js';
+import { S, save } from '../../../shared/state.js';
 import { WEAPONS } from '../../weaponGeneration/data/weapons.js';
 import { WEAPON_ICONS } from '../../weaponGeneration/data/weaponIcons.js';
 import { equipItem, unequip, removeFromInventory } from '../mutators.js';
