+++ src/features/proficiency/ui/weaponProficiencyDisplay.js
@@ -1,10 +1,10 @@
-import { S } from "../../../game/state.js";
+import { S } from "../../../shared/state.js";
 import { on } from "../../../shared/events.js";
 import { getEquippedWeapon } from "../../inventory/selectors.js";
 import { getProficiency } from "../selectors.js";
 import { WEAPON_TYPES } from "../../weaponGeneration/data/weaponTypes.js";
 import { WEAPON_ICONS } from "../../weaponGeneration/data/weaponIcons.js";
-import { setText, setFill } from "../../../game/utils.js";
+import { setText, setFill } from "../../../shared/utils/dom.js";
 
 export function updateWeaponProficiencyDisplay(state = S) {
   const weapon = getEquippedWeapon(state);
