+++ src/shared/mutators.js
@@ -1,0 +1,9 @@
+export * as ability from '../features/ability/mutators.js';
+export * as adventure from '../features/adventure/mutators.js';
+export * as affixes from '../features/affixes/mutators.js';
+export * as combat from '../features/combat/mutators.js';
+export * as inventory from '../features/inventory/mutators.js';
+export * as loot from '../features/loot/mutators.js';
+export * as proficiency from '../features/proficiency/mutators.js';
+export * as progression from '../features/progression/mutators.js';
+export * as weaponGeneration from '../features/weaponGeneration/mutators.js';
