+++ src/shared/selectors.js
@@ -1,0 +1,9 @@
+export * as ability from '../features/ability/selectors.js';
+export * as adventure from '../features/adventure/selectors.js';
+export * as affixes from '../features/affixes/selectors.js';
+export * as combat from '../features/combat/selectors.js';
+export * as inventory from '../features/inventory/selectors.js';
+export * as loot from '../features/loot/selectors.js';
+export * as proficiency from '../features/proficiency/selectors.js';
+export * as progression from '../features/progression/selectors.js';
+export * as weaponGeneration from '../features/weaponGeneration/selectors.js';
